refactor(web/listings): split OfferProductPickerModal into row/rail sub-components
The picker had grown to ~1036 lines after the unify-publish work, past the
point flagged as tech debt in #1819. Extracted the per-product/variant row
into offer-product-picker-row.tsx, the selection review rail into
offer-product-picker-rail.tsx, shared types into offer-product-picker.types.ts,
and pure derivation logic into lib/offer-product-picker-selectors.ts (plus a
new resolvePublishDestination helper in the existing lib/publish-destinations.ts).
No behaviour change - the modal orchestrator keeps the same public API and
all state/callbacks; existing tests pass unmodified.

Closes #1819
